making the ROMSEARCH functional - at least as far as display is concenren. Now I need to figure out how to dynamiclaly allocate memory to store the addresses that I find.
Also, the subroutinges creting the signal timing should be replaced with assembly code
without interupts to make tsure the timing is what I want it to be.
This should more or less wrap up the MASTER part of the 1-wire communication. A next prohject
will be to write the pieces that alow the uc to be a slave in another 1-wire channel.